+++ lib/pmdtester/builders/pmd_report_builder.rb
@@ -25,7 +25,7 @@ module PmdTester
     def get_pmd_binary_file
       logger.info "#{@pmd_branch_name}: Start packaging PMD"
       Dir.chdir(@local_git_repo) do
-        build_branch_sha = Cmd.execute("git rev-parse #{@pmd_branch_name}^{commit}")
+        build_branch_sha = Cmd.execute_successfully("git rev-parse #{@pmd_branch_name}^{commit}")
 
         checkout_build_branch # needs a clean working tree, otherwise fails
 
@@ -70,28 +70,28 @@ module PmdTester
                       ' -Dcheckstyle.skip=true' \
                       ' -Dpmd.skip=true' \
                       ' -T1C -B'
-        Cmd.execute(package_cmd)
+        Cmd.execute_successfully(package_cmd)
       end
 
       logger.info "#{@pmd_branch_name}: Extracting the zip"
-      Cmd.execute("unzip -qo #{pmd_dist_target} -d pmd-dist/target/exploded")
-      Cmd.execute("mv pmd-dist/target/exploded/pmd-bin-#{@pmd_version} #{into_dir}")
+      Cmd.execute_successfully("unzip -qo #{pmd_dist_target} -d pmd-dist/target/exploded")
+      Cmd.execute_successfully("mv pmd-dist/target/exploded/pmd-bin-#{@pmd_version} #{into_dir}")
     end
 
     def determine_pmd_version
       version_cmd = "./mvnw -q -Dexec.executable=\"echo\" -Dexec.args='${project.version}' " \
                     '--non-recursive org.codehaus.mojo:exec-maven-plugin:1.5.0:exec'
-      Cmd.execute(version_cmd)
+      Cmd.execute_successfully(version_cmd)
     end
 
     def get_last_commit_sha
       get_last_commit_sha_cmd = 'git rev-parse HEAD^{commit}'
-      Cmd.execute(get_last_commit_sha_cmd)
+      Cmd.execute_successfully(get_last_commit_sha_cmd)
     end
 
     def get_last_commit_message
       get_last_commit_message_cmd = 'git log -1 --pretty=%B'
-      Cmd.execute(get_last_commit_message_cmd)
+      Cmd.execute_successfully(get_last_commit_message_cmd)
     end
 
     def generate_pmd_report(project)
@@ -107,14 +107,16 @@ module PmdTester
                 "#{auxclasspath_option}" \
                 "#{pmd7? ? ' --no-progress' : ''}"
       start_time = Time.now
+      exit_code = nil
       if File.exist?(project.get_pmd_report_path(@pmd_branch_name))
         logger.warn "#{@pmd_branch_name}: Skipping PMD run - report " \
                     "#{project.get_pmd_report_path(@pmd_branch_name)} already exists"
       else
-        Cmd.execute(pmd_cmd)
+        status = Cmd.execute(pmd_cmd, project.get_project_target_dir(@pmd_branch_name))
+        exit_code = status.exitstatus
       end
       end_time = Time.now
-      [end_time - start_time, end_time]
+      [end_time - start_time, end_time, exit_code]
     end
 
     def generate_config_for(project)
@@ -141,12 +143,12 @@ module PmdTester
         progress_logger = SimpleProgressLogger.new("generating #{project.name}'s PMD report")
         progress_logger.start
         generate_config_for(project)
-        execution_time, end_time = generate_pmd_report(project)
+        execution_time, end_time, exit_code = generate_pmd_report(project)
         progress_logger.stop
         sum_time += execution_time
 
-        report_details = PmdReportDetail.new(execution_time: execution_time, timestamp: end_time)
-        report_details.save(project.get_report_info_path(@pmd_branch_name))
+        PmdReportDetail.create(execution_time: execution_time, timestamp: end_time,
+                               exit_code: exit_code, report_info_path: project.get_report_info_path(@pmd_branch_name))
         logger.info "#{project.name}'s PMD report was generated successfully"
       end
 
@@ -169,7 +171,7 @@ module PmdTester
     def checkout_build_branch
       logger.info "#{@pmd_branch_name}: Checking out the branch"
       # note that this would fail if the tree is dirty
-      Cmd.execute("git checkout #{@pmd_branch_name}")
+      Cmd.execute_successfully("git checkout #{@pmd_branch_name}")
 
       # determine the version
       @pmd_version = determine_pmd_version
@@ -195,7 +197,7 @@ module PmdTester
     end
 
     def wd_has_dirty_git_changes
-      !Cmd.execute('git status --porcelain').empty?
+      !Cmd.execute_successfully('git status --porcelain').empty?
     end
 
     def should_use_long_cli_options