+++ lib/pmdtester/report_diff.rb
@@ -57,17 +57,22 @@ module PmdTester
                 :configerrors_by_rule,
                 :exec_time,
                 :timestamp,
+                :exit_code,
                 :file
+
+    attr_accessor :report_folder
 
     def initialize(report_document: nil,
                    file: '',
                    exec_time: 0,
-                   timestamp: '0')
+                   timestamp: '0',
+                   exit_code: '?')
       initialize_empty
       initialize_with_report_document report_document unless report_document.nil?
       @exec_time = exec_time
       @timestamp = timestamp
       @file = file
+      @exit_code = exit_code
     end
 
     def self.empty
@@ -108,6 +113,7 @@ module PmdTester
       @violations_by_file = CollectionByFile.new
       @errors_by_file = CollectionByFile.new
       @configerrors_by_rule = {}
+      @report_folder = ''
     end
   end