+++ test/test_cmd.rb
@@ -6,28 +6,48 @@ require 'test_helper'
 class TestCmd < Test::Unit::TestCase
   include PmdTester
 
-  def test_get_stdout
-    stdout = Cmd.execute('echo Hello, World!')
-    assert_equal('Hello, World!', stdout)
+  def setup
+    @tempdir = 'test-TestCmd-temp'
+    Dir.mkdir @tempdir unless Dir.exist?(@tempdir)
   end
 
-  def test_invalid_cmd(cmd)
-    expected_msg = "#{CmdException::COMMON_MSG} '#{cmd}'"
-    begin
-      Cmd.execute(cmd)
-    rescue CmdException => e
-      assert_equal(cmd, e.cmd)
-      assert_equal(expected_msg, e.message)
-    end
+  def teardown
+    Dir.each_child(@tempdir) { |x| File.unlink("#{@tempdir}/#{x}") }
+    Dir.rmdir(@tempdir)
+  end
+
+  def test_get_stdout
+    stdout = Cmd.execute_successfully('echo Hello, World!')
+    assert_equal('Hello, World!', stdout)
   end
 
   def test_invalid_cmd_1
     cmd = 'cd DIR_NO_EXIST'
-    test_invalid_cmd(cmd)
+    run_invalid_cmd(cmd)
   end
 
   def test_invalid_cmd_2
     cmd = 'false'
-    test_invalid_cmd(cmd)
+    run_invalid_cmd(cmd)
+  end
+
+  def test_failing_cmd
+    status = Cmd.execute('echo Hello; echo World >&2; exit 5', @tempdir)
+
+    assert_equal("Hello\n", File.read("#{@tempdir}/stdout.txt"))
+    assert_equal("World\n", File.read("#{@tempdir}/stderr.txt"))
+    assert_equal(5, status.exitstatus)
+  end
+
+  private
+
+  def run_invalid_cmd(cmd)
+    expected_msg = "#{CmdException::COMMON_MSG} '#{cmd}'"
+    begin
+      Cmd.execute_successfully(cmd)
+    rescue CmdException => e
+      assert_equal(cmd, e.cmd)
+      assert(e.message.start_with?(expected_msg))
+    end
   end
 end